fixed claiming submissions and deleted redundant models
