+++ app/assets/javascripts/aker.js
@@ -3,7 +3,7 @@
 
   self.Aker.claim = function(submissionIds, collectionId) {
     $.post({
-      url: '/material_submissions/claim',
+      url: '/claim_submissions/claim',
       data: {
         submission_ids: submissionIds,
         collection_id: collectionId
