+++ app/controllers/material_receptions_controller.rb
@@ -13,7 +13,7 @@ class MaterialReceptionsController < ApplicationController
     if @material_reception.save
       ReceptionMailer.material_reception(@material_reception).deliver_later
       # Only check if the set is complete is material_reception saved
-      if @material_reception.complete_set?
+      if @material_reception.all_received?
         ReceptionMailer.complete_set(@material_reception).deliver_later
       end
     end
@@ -27,7 +27,7 @@ class MaterialReceptionsController < ApplicationController
 
   def set_labware
     @labware = Labware.with_barcode(material_reception_params[:barcode_value]).first
-    @labware_id = @labware&.uuid
+    @labware_id = @labware&.id
   end
 
 end
