--- app/models/barcode.rb
@@ -1,3 +1,0 @@
-class Barcode < ApplicationRecord
-  belongs_to :barcodeable, polymorphic: true
-end
