+++ app/models/material_reception.rb
@@ -1,15 +1,7 @@
 class MaterialReception < ApplicationRecord
-  #belongs_to :labware
-
-  before_create :receive_labware
 
   validates :labware_id, uniqueness: { message: "cannot be received twice" }
   validate :validate_barcode_printed, on: :create
-
-  def receive_labware
-    return false unless labware && labware.barcode_printed?
-    labware.received_unclaimed
-  end
 
   def barcode_value
     labware&.barcode
@@ -20,11 +12,10 @@ class MaterialReception < ApplicationRecord
   end
 
   def barcode_value=(barcode)
-    self.labware_id = Labware.with_barcode(barcode).first.uuid # TODO - id or uuid?
+    self.labware_id = Labware.with_barcode(barcode).first.id
   end
 
   def validate_barcode_printed
-    debugger
     unless labware
       errors.add(:labware, "could not be found")
     else
@@ -33,14 +24,14 @@ class MaterialReception < ApplicationRecord
   end
 
   def labware_already_received?
-    MaterialReception.where(:labware_id => labware.uuid).count > 0
+    MaterialReception.where(:labware_id => labware.id).count > 0
   end
 
 
   def presenter
     if invalid?
       return {:error => 'Cannot find the barcode'} unless barcode_value
-      return {:error => 'Labware already received'} if labware_already_received?    
+      return {:error => 'Labware already received'} if labware_already_received?
       return {:error => 'This barcode has not been printed yet. Please contact the administrator'} unless labware.barcode_printed?
     else
       {
@@ -51,8 +42,8 @@ class MaterialReception < ApplicationRecord
     end
   end
 
-  def complete_set?
-    labware.material_submission.labwares.all?(&:received_unclaimed?)
+  def all_received?
+    labware.material_submission.labwares.all?(&:recieved)
   end
 
 end
