+++ db/migrate/20170426140233_drop_old_tables.rb
@@ -4,5 +4,6 @@ class DropOldTables < ActiveRecord::Migration[5.0]
     drop_table :biomaterials
     drop_table :labwares
     drop_table :material_submission_labwares
+    drop_table :barcodes
   end
 end
